+++ __tests__/unit/defaults.test.js
@@ -55,123 +55,10 @@ describe("defaults", () => {
         });
     });
 
-    describe("packages", () => {
-        const testGetPackageDataName = "My Package";
-        const testGetPackageOrdNamespace = "customer.sample";
-        var appConfig = {};
-        it("should return default value if policyLevels contains sap", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                policyLevels: ["sap:policy"],
-                hasSAPPolicyLevel: true,
-            };
-            expect(defaults.packages(appConfig)).toMatchSnapshot();
-        });
-
-        it("should return default value if policyLevels does not contain sap", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                policyLevels: ["policy"],
-            };
-            expect(defaults.packages(appConfig)).toMatchSnapshot();
-        });
-
-        it("should return custom value if user defined in .cdsrc.json", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                env: {
-                    packages: [
-                        {
-                            vendor: "sap:vendor:SAP:",
-                            tags: ["custom"],
-                        },
-                    ],
-                },
-                policyLevels: ["policy"],
-            };
-            expect(
-                defaults.packages(appConfig, [{ ordId: "customer:product:eb.bm.tests:", vendor: "sap:vendor:SAP:" }]),
-            ).toMatchSnapshot();
-        });
-
-        it("should return only custom value if user definitions in .cdsrc.json are done correctly", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                env: {
-                    packages: [
-                        {
-                            licenseType: "important license", // valid type
-                            runtimeRestriction: 4, // incorrect type
-                            industry: "Finance", // incorrect type
-                            labels: ["correct"], // valid type
-                            invalidKey: "should be removed", // invalid key
-                        },
-                    ],
-                },
-                policyLevels: ["policy"],
-            };
-            expect(
-                defaults.packages(appConfig, [{ ordId: "customer:product:eb.bm.tests:", vendor: "sap:vendor:SAP:" }]),
-            ).toMatchSnapshot();
-        });
-
-        it("should use existingProductId if provided in .cdsrc.json", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                existingProductORDId: "sap:product:SAPServiceCloudV2:",
-                policyLevels: ["policy"],
-            };
-            expect(defaults.packages(appConfig)).toMatchSnapshot();
-        });
-
-        it("should use existingProductId if existingProductId and custom product both provided in .cdsrc.json", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                existingProductORDId: "sap:product:SAPServiceCloudV2:",
-                policyLevels: ["policy"],
-            };
-            expect(
-                defaults.packages(appConfig, [{ ordId: "customer:product:eb.bm.tests:", vendor: "sap:vendor:SAP:" }]),
-            ).toMatchSnapshot();
-        });
-
-        it("should use custom vendor if it defined in .cdsrc.json", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                env: {
-                    packages: [
-                        {
-                            vendor: "sap:vendor:SAP:",
-                        },
-                    ],
-                },
-                policyLevels: ["policy"],
-            };
-            expect(defaults.packages(appConfig)).toMatchSnapshot();
-        });
-
-        it("should not contain partOfProducts if no productsOrdId found", () => {
-            appConfig = {
-                appName: testGetPackageDataName,
-                ordNamespace: testGetPackageOrdNamespace,
-                policyLevels: ["policy"],
-            };
-            expect(defaults.packages(appConfig)).toMatchSnapshot();
-        });
-    });
-
     describe("baseTemplate", () => {
-
         beforeEach(() => {
             slice.mockClear();
-        })
+        });
 
         it("should return correct value when no tenant is given and toggles & extensions are disabled", () => {
             cds.env.requires.toggles = false;
