+++ lib/configuration.js
@@ -7,6 +7,7 @@ const defaults = require("./defaults");
 const { getRFC3339Date } = require("./common/utils");
 const { createAuthConfig } = require("./auth/authentication");
 const { CDS_ELEMENT_KIND, ORD_EXTENSIONS_PREFIX } = require("./constants");
+const { ensureAccessStrategies, getAccessStrategiesFromAuthConfig } = require("./access-strategies");
 
 module.exports = class Configuration {
     constructor(csn) {
@@ -34,6 +35,7 @@ module.exports = class Configuration {
         this._existingProductORDId = env?.existingProductORDId;
         this._appName = packageName.replace(/^@/, "").replace(/[@/]/g, "-");
         this._policyLevels = env?.policyLevels || (env?.policyLevel && [env?.policyLevel]) || defaults.policyLevels;
+        this._accessStrategies = ensureAccessStrategies(getAccessStrategiesFromAuthConfig(authConfig.accessStrategies ?? []));
     }
 
     get env() {
@@ -66,6 +68,10 @@ module.exports = class Configuration {
 
     get policyLevels() {
         return this._policyLevels;
+    }
+
+    get accessStrategies() {
+        return this._accessStrategies;
     }
 
     get internalNamespace() {
