+++ lib/ord.js
@@ -3,6 +3,7 @@ const cds = require("@sap/cds");
 const defaults = require("./defaults");
 const Configuration = require("./configuration");
 const { createGroups } = require("./templates/group");
+const { createPackages } = require("./templates/package");
 const { createProducts } = require("./templates/product");
 const { createEntityTypes } = require("./templates/entity-type");
 const { createAPIResources } = require("./templates/api-resource");
@@ -33,20 +34,19 @@ const _filterUnusedPackages = (ordDocument) => {
 
 module.exports = (csn, extensions = []) => {
     const appConfig = new Configuration(cds.linked(csn));
-    const products = createProducts(appConfig);
     const ordDocument = [...(extensions || []), getCustomORDContent(appConfig)]
         .filter((extension) => !!extension)
         .reduce((document, extension) => compareAndHandleCustomORDContentWithExistingContent(document, extension), {
-            products: products,
             $schema: defaults.$schema,
             groups: createGroups(appConfig),
+            products: createProducts(appConfig),
+            packages: createPackages(appConfig),
             policyLevels: appConfig.policyLevels,
             entityTypes: createEntityTypes(appConfig),
             apiResources: createAPIResources(appConfig),
             eventResources: createEventResources(appConfig),
-            packages: defaults.packages(appConfig, products),
             consumptionBundles: appConfig.env?.consumptionBundles ?? [],
-            description: appConfig.env?.description || defaults.description,
+            description: appConfig.env?.description ?? defaults.description,
             integrationDependencies: createIntegrationDependencies(appConfig),
             openResourceDiscovery: appConfig.env?.openResourceDiscovery || defaults.openResourceDiscovery,
         });
