+++ lib/templates/api-resource.js
@@ -1,9 +1,14 @@
 const Logger = require("../logger");
 const defaults = require("../defaults");
+const { createPackages } = require("./package");
 const { resolveApiResourceProtocol } = require("../protocol-resolver");
-const { readORDExtensions, isPrimaryDataProductService, isValidService, isEventsOnlyService } = require("../common/utils");
 const { resolveVisibility, resolveServiceName, flattenEntityGraph } = require("../common/utils");
-const { ensureAccessStrategies, getAccessStrategiesFromAuthConfig } = require("../access-strategies");
+const {
+    readORDExtensions,
+    isPrimaryDataProductService,
+    isValidService,
+    isEventsOnlyService,
+} = require("../common/utils");
 const {
     DESCRIPTION_PREFIX,
     ORD_RESOURCE_TYPE,
@@ -57,18 +62,11 @@ function _getExposedEntityTypes(service) {
  * @private
  */
 function _getResourceDefinition(resourceType, mediaType, ordId, serviceName, fileExtension, accessStrategies) {
-    // Validate and ensure access strategies are present
-    // In non-strict mode, this will log error and fallback to 'open'
-    // In strict mode (cds.env.ord.strictAccessStrategies = true), this will throw
-    const validatedStrategies = ensureAccessStrategies(accessStrategies, {
-        resourceName: `${serviceName} (${resourceType})`,
-    });
-
     return {
         type: resourceType,
         mediaType: `application/${mediaType}`,
         url: `/ord/v1/${ordId}/${serviceName}.${fileExtension}`,
-        accessStrategies: validatedStrategies,
+        accessStrategies: accessStrategies,
     };
 }
 
@@ -83,11 +81,11 @@ function _getResourceDefinition(resourceType, mediaType, ordId, serviceName, fil
 const createAPIResourceTemplate = (srvDefinition, appConfig) => {
     const ordExtensions = readORDExtensions(srvDefinition);
     const visibility = resolveVisibility(appConfig, srvDefinition);
-    const packageIds = defaults.packages(appConfig)?.map((pkg) => pkg.ordId) || [];
+    const packageIds = createPackages(appConfig)?.map((pkg) => pkg.ordId) || [];
     const packageId = _getPackageID(appConfig.ordNamespace, packageIds, ORD_RESOURCE_TYPE.api, visibility);
     const protocolResults = resolveApiResourceProtocol(srvDefinition);
     const apiResources = [];
-    const accessStrategies = getAccessStrategiesFromAuthConfig(appConfig.authConfig.accessStrategies || []);
+    const accessStrategies = appConfig.accessStrategies;
 
     // If no protocols were generated, skip this service
     if (protocolResults.length === 0) {
@@ -166,9 +164,7 @@ const createAPIResourceTemplate = (srvDefinition, appConfig) => {
                         type: "graphql-sdl",
                         mediaType: "text/plain",
                         url: `/ord/v1/${ordId}/${srvDefinition.name}.graphql`,
-                        accessStrategies: ensureAccessStrategies(accessStrategies, {
-                            resourceName: `${srvDefinition.name} (graphql-sdl)`,
-                        }),
+                        accessStrategies: accessStrategies,
                     },
                 ];
             } else if (apiProtocol === ORD_API_PROTOCOL.ODATA_V2) {
