+++ lib/templates/entity-type.js
@@ -1,5 +1,6 @@
 const defaults = require("../defaults");
 const Logger = require("../logger");
+const { createPackages } = require("./package");
 const { readORDExtensions, isExposedEntityType } = require("../common/utils");
 const {
     LEVEL,
@@ -53,7 +54,7 @@ const RESOLVERS = Object.freeze({
     partOfPackage: (entity, appConfig) => {
         const visibility = RESOLVERS.visibility(entity, appConfig);
         const name = appConfig.appName?.replace(/[^a-zA-Z0-9]/g, "");
-        const packages = defaults.packages(appConfig).map((pkg) => pkg.ordId);
+        const packages = createPackages(appConfig).map((pkg) => pkg.ordId);
         const suffix = visibility === RESOURCE_VISIBILITY.public ? "" : `-${visibility}`;
 
         return (
