+++ lib/templates/event-resource.js
@@ -1,6 +1,6 @@
 const defaults = require("../defaults");
 const entityTypeTemplate = require("./entity-type");
-const { ensureAccessStrategies, getAccessStrategiesFromAuthConfig } = require("../access-strategies");
+const { createPackages } = require("./package");
 const {
     readORDExtensions,
     resolveVisibility,
@@ -74,7 +74,7 @@ const RESOLVERS = Object.freeze({
     partOfPackage: (service, appConfig) => {
         const visibility = RESOLVERS.visibility(service, appConfig);
         const name = appConfig.appName?.replace(/[^a-zA-Z0-9]/g, "");
-        const packages = defaults.packages(appConfig).map((pkg) => pkg.ordId);
+        const packages = createPackages(appConfig).map((pkg) => pkg.ordId);
         const suffix = visibility === RESOURCE_VISIBILITY.public ? "" : `-${visibility}`;
 
         return (
@@ -88,16 +88,14 @@ const RESOLVERS = Object.freeze({
     resourceDefinitions: (service, appConfig) => {
         const name = service.name;
         const ordId = RESOLVERS.ordId(service, appConfig);
-        const accessStrategies = getAccessStrategiesFromAuthConfig(appConfig.authConfig.accessStrategies);
+        const accessStrategies = appConfig.accessStrategies;
 
         return [
             {
                 type: "asyncapi-v2",
                 mediaType: `application/json`,
                 url: `/ord/v1/${ordId}/${name}.asyncapi2.json`,
-                accessStrategies: ensureAccessStrategies(accessStrategies, {
-                    resourceName: `${name} (asyncapi-v2)`,
-                }),
+                accessStrategies: accessStrategies,
             },
         ];
     },
