+++ lib/templates/integration-dependency.js
@@ -1,4 +1,5 @@
 const defaults = require("../defaults");
+const { createPackages } = require("./package");
 const { readORDExtensions, isExternalDataProduct } = require("../common/utils");
 const { RESOURCE_VISIBILITY, EXTERNAL_DP_ORD_ID_ANNOTATION } = require("../constants");
 
@@ -39,7 +40,7 @@ const RESOLVERS = Object.freeze({
     partOfPackage: (appConfig) => {
         const visibility = RESOLVERS.visibility(appConfig);
         const name = appConfig.appName?.replace(/[^a-zA-Z0-9]/g, "");
-        const packages = defaults.packages(appConfig).map((pkg) => pkg.ordId);
+        const packages = createPackages(appConfig).map((pkg) => pkg.ordId);
         const suffix = visibility === RESOURCE_VISIBILITY.public ? "" : `-${visibility}`;
 
         return (
