+++ lib/templates/product.js
@@ -17,9 +17,9 @@ const nameWithSpaces = (name) => {
 function createProduct(appConfig) {
     return {
         vendor: "customer:vendor:Customer:",
-        title: nameWithSpaces(appConfig.appName),
-        ordId: `customer:product:${nameWithDot(appConfig.appName)}:`,
-        shortDescription: `Short description of ${nameWithSpaces(appConfig.appName)}`,
+        title: nameWithSpaces(appConfig.packageName),
+        ordId: `customer:product:${nameWithDot(appConfig.packageName)}:`,
+        shortDescription: `Short description of ${nameWithSpaces(appConfig.packageName)}`,
     };
 }
